+++ feedback-service/src/main/java/org/alexreverse/controller/PostReviewsRestController.java
@@ -20,7 +20,7 @@ public class PostReviewsRestController {
     private final PostReviewsService postReviewsService;
 
     @GetMapping("by-post-id/{postId:\\d+}")
-    public Flux<PostReview> findPostReviewsByPostId(@PathVariable("postId") int postId) {
+    public Flux<PostReview> findPostReviewsByPostId(@PathVariable("postId") Long postId) {
         return this.postReviewsService.findPostReviewsByPost(postId);
     }
