+++ feedback-service/src/main/java/org/alexreverse/controller/payload/NewFavouritePostPayload.java
@@ -3,7 +3,7 @@ package org.alexreverse.controller.payload;
 import jakarta.validation.constraints.NotNull;
 
 public record NewFavouritePostPayload(@NotNull(message = "{feedback.posts.favourites.create.errors.post_id_is_null}")
-                                      Integer postId,
+                                      Long postId,
                                       @NotNull
                                       String userName) {
 }
