+++ feedback-service/src/main/java/org/alexreverse/repository/PostRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Flux;
 
-public interface PostRepository extends ReactiveCrudRepository<Post, Integer> {
+public interface PostRepository extends ReactiveCrudRepository<Post, Long> {
 
     @Query(value = "select p from Post p where p.title ilike :filter")
     Flux<Post> findAllByTitleLikeIgnoreCase(@Param("filter") String filter);
